pyproject.toml
[project]
name = "nonebot-plugin-bilichat"

version = "5.3.2"

description = "多种B站链接解析，视频词云，AI总结，你想要的都在 bilichat"
authors = [
    { name = "djkcyl", email = "cyl@cyllive.cn" },
    { name = "Well404", email = "well_404@outlook.com" },
]
dependencies = [
    "bilireq>=0.2.8",
    "qrcode>=7.4.2",
    "pillow>=9.5.0",
    "lxml>=4.9.2",
    "nonebot-plugin-localstore>=0.4.1",
    "nonebot2[fastapi,websockets]>=2.0.0",
    "httpx>=0.24.1",
    "dynrender-skia-opt>=0.3.8",
    "nonebot-plugin-apscheduler>=0.3.0",
    "nonebot-plugin-send-anything-anywhere>=0.4.0",
    "packaging>=23.2",
    "python-multipart>=0.0.6",
    "nonebot-plugin-alconna>=0.41.1",
]
requires-python = ">=3.8,<4.0"
readme = "README.md"
license = { text = "AGPL3.0" }
[project.optional-dependencies]
extra = [
    "numpy>=1.20.1,<1.25.0",
    "nonebot-plugin-sentry>=0.2.2",
    "nonebot-plugin-htmlrender>=0.2.0.3",
    "nonebot-plugin-mongodb>=0.1.0",
]
wordcloud = ["jieba>=0.42.1", "wordcloud>=1.8.2.2"]
summary = [
    "tiktoken>=0.6.0",
]
all = [
    "numpy>=1.20.1,<1.25.0",
    "jieba>=0.42.1",
    "wordcloud>=1.8.2.2",
    "nonebot-plugin-htmlrender>=0.2.0.3",
    "nonebot-plugin-mongodb>=0.1.0",
    "nonebot-plugin-sentry>=0.4.1",
    "tiktoken>=0.6.0",
]

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
    "nonebot-adapter-onebot>=2.2.4",
    "nonebot-adapter-mirai2>=0.0.22",
    "nonebot-adapter-qq>=1.0.1",
    "viztracer>=0.16.1",
    "nb-cli>=1.2.8",
    "nonebot-adapter-satori>=0.10.3",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
